Updated Java requirements and links
This is part one of a bigger update. This part includes:
* links to newer Java versions (Java 1.5 is old now);
* links to the schematron files (skeleton etc.) on www.schematron.com
were outdated (Error 404) and now point to copies of the same old files
in the bentoweb-tcdl repository.

In the second part of the updated, the old ISO schematron files will
(hopefully) be replaced with the current ones from www.schematron.com.